requirements.txt
pyyaml >= 5.4.1
feedgen >= 0.9.0